+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# 2.0.0
+
+* [BC Break]: Configuration of the serializer generator changed to configuration model.
+  The new format allows to more precisely specify which serializers to generate.
+
 # 1.0.0
 
 Initial release